+++ NEWS.md
@@ -12,6 +12,7 @@ this is fixed thanks to Christophe Dervieux and Nick Bart. Repeating
 the caption along pages can be desactivated with command 
 `set_table_properties(opts_pdf = list(caption_repeat = FALSE))`.
 - `as_flextable()` now works on tabular objects and `R < 4.1`.
+- `to_html(type = "img")` now use the correct width and height
 
 ## Changes